dcap client: close tunnel when closing control channel
The dcap client includes the concept of optionally encoding the control 
channel.  This is enabled through loading a DSO plugin.  In addition to 
the encoding and decoding functions, the plugins also have an 
initialise and destroy functions.

The dcap code calls the initialise method but fails to ever call the 
destroy function of the plugin.  Under certain circumstances this can 
lead to a memory leak.

This patch updates the dcap client so it calls the tunnel's destroy 
function (if a tunnel is being used) when the control channel is 
closed.

FOR RELEASE NOTES:

Mention that a memory leak for gsidcap under rare conditions has been 
fixed.

Target: trunk
Require-notes: yes
Require-book: no
Patch: http://rb.dcache.org/r/5043/
Acked-by: Tigran Mkrtchyan
Require-notes: yes
Require-book: no